refractor(views): seccionar views.py en archivos admin.py, client.py y auth.py
