+++ users/views/__init__.py
@@ -1,0 +1,8 @@
+# Importar vistas de autenticación
+from .auth_views import *
+
+# Importar vistas de cliente
+from .client_views import *
+
+# Importar vistas de administrador
+from .admin_views import *
